SASS-2406 - Unit test page content rather than in integration tests in accommodation and assets packges.
