+++ app/controllers/benefits/assets/AssetsTransfersBenefitsAmountController.scala
@@ -19,12 +19,12 @@ package controllers.benefits.assets
 import actions.AuthorisedAction
 import config.{AppConfig, ErrorHandler}
 import controllers.employment.routes.CheckYourBenefitsController
-import forms.{AmountForm, FormUtils}
+import forms.FormUtils
+import forms.benefits.assets.AssetsFormsProvider
 import models.AuthorisationRequest
 import models.employment.EmploymentBenefitsType
 import models.mongo.{EmploymentCYAModel, EmploymentUserData}
 import models.redirects.ConditionalRedirect
-import play.api.data.Form
 import play.api.i18n.I18nSupport
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents, Result}
 import services.RedirectService.{assetTransferAmountRedirects, redirectBasedOnCurrentAnswers}
@@ -42,8 +42,9 @@ class AssetsTransfersBenefitsAmountController @Inject()(authAction: AuthorisedAc
                                                         pageView: AssetsTransfersBenefitsAmountView,
                                                         employmentSessionService: EmploymentSessionService,
                                                         assetsService: AssetsService,
-                                                        errorHandler: ErrorHandler)
-                                                        (implicit val cc: MessagesControllerComponents, appConfig: AppConfig, ec: ExecutionContext)
+                                                        errorHandler: ErrorHandler,
+                                                        formsProvider: AssetsFormsProvider)
+                                                       (implicit cc: MessagesControllerComponents, appConfig: AppConfig, ec: ExecutionContext)
   extends FrontendController(cc) with I18nSupport with SessionHelper with FormUtils {
 
   def show(taxYear: Int, employmentId: String): Action[AnyContent] = authAction.async { implicit request =>
@@ -53,7 +54,7 @@ class AssetsTransfersBenefitsAmountController @Inject()(authAction: AuthorisedAc
         redirectBasedOnCurrentAnswers(taxYear, employmentId, optCya, EmploymentBenefitsType)(redirects(_, taxYear, employmentId)) { cya =>
           val cyaAmount = cya.employment.employmentBenefits.flatMap(_.assetsModel.flatMap(_.assetTransfer))
 
-          val form = fillFormFromPriorAndCYA(amountForm(request.user.isAgent), prior, cyaAmount, employmentId)(
+          val form = fillFormFromPriorAndCYA(formsProvider.assetTransfersAmountForm(request.user.isAgent), prior, cyaAmount, employmentId)(
             employment => employment.employmentBenefits.flatMap(_.benefits.flatMap(_.assetTransfer))
           )
           Future.successful(Ok(pageView(taxYear, form, cyaAmount, employmentId)))
@@ -67,7 +68,7 @@ class AssetsTransfersBenefitsAmountController @Inject()(authAction: AuthorisedAc
       employmentSessionService.getSessionDataResult(taxYear, employmentId) { cya =>
 
         redirectBasedOnCurrentAnswers(taxYear, employmentId, cya, EmploymentBenefitsType)(redirects(_, taxYear, employmentId)) { cya =>
-          amountForm(request.user.isAgent).bindFromRequest().fold(
+          formsProvider.assetTransfersAmountForm(request.user.isAgent).bindFromRequest().fold(
             formWithErrors => {
               val cyaAmount = cya.employment.employmentBenefits.flatMap(_.assetsModel.flatMap(_.assetTransfer))
               Future.successful(BadRequest(pageView(taxYear, formWithErrors, cyaAmount, employmentId)))
@@ -88,12 +89,6 @@ class AssetsTransfersBenefitsAmountController @Inject()(authAction: AuthorisedAc
         RedirectService.benefitsSubmitRedirect(employmentUserData.employment, nextPage)(taxYear, employmentId)
     }
   }
-
-  private def amountForm(isAgent: Boolean): Form[BigDecimal] = AmountForm.amountForm(
-    emptyFieldKey = s"benefits.assetTransfersAmount.error.noEntry.${if (isAgent) "agent" else "individual"}",
-    wrongFormatKey = s"benefits.assetTransfersAmount.error.incorrectFormat.${if (isAgent) "agent" else "individual"}",
-    exceedsMaxAmountKey = s"benefits.assetTransfersAmount.error.overMaximum.${if (isAgent) "agent" else "individual"}"
-  )
 
   private def redirects(cya: EmploymentCYAModel, taxYear: Int, employmentId: String): Seq[ConditionalRedirect] = {
     assetTransferAmountRedirects(cya, taxYear, employmentId)
