+++ app/views/benefits/accommodation/LivingAccommodationBenefitsAmountView.scala.html
