+++ it/controllers/benefits/accommodation/LivingAccommodationBenefitsControllerISpec.scala
@@ -18,8 +18,6 @@ package controllers.benefits.accommodation
 
 import forms.YesNoForm
 import models.benefits.AccommodationRelocationModel
-import org.jsoup.Jsoup
-import org.jsoup.nodes.Document
 import play.api.http.HeaderNames
 import play.api.http.Status.{BAD_REQUEST, OK, SEE_OTHER}
 import play.api.libs.ws.WSResponse
@@ -29,214 +27,28 @@ import support.builders.models.benefits.AccommodationRelocationModelBuilder.anAc
 import support.builders.models.benefits.BenefitsViewModelBuilder.aBenefitsViewModel
 import support.builders.models.mongo.EmploymentCYAModelBuilder.anEmploymentCYAModel
 import support.builders.models.mongo.EmploymentUserDataBuilder.{anEmploymentUserData, anEmploymentUserDataWithBenefits}
-import utils.PageUrls.{accommodationRelocationBenefitsUrl, checkYourBenefitsUrl, fullUrl, livingAccommodationBenefitsAmountUrl,
-  livingAccommodationBenefitsUrl, overviewUrl, qualifyingRelocationBenefitsUrl, travelOrEntertainmentBenefitsUrl}
+import utils.PageUrls.{accommodationRelocationBenefitsUrl, checkYourBenefitsUrl, fullUrl, livingAccommodationBenefitsAmountUrl, livingAccommodationBenefitsUrl, overviewUrl, qualifyingRelocationBenefitsUrl, travelOrEntertainmentBenefitsUrl}
 import utils.{EmploymentDatabaseHelper, IntegrationTest, ViewHelpers}
 
 class LivingAccommodationBenefitsControllerISpec extends IntegrationTest with ViewHelpers with EmploymentDatabaseHelper {
 
   private val employmentId: String = "employmentId"
 
-  object Selectors {
-    val paragraphSelector = "#main-content > div > div > p"
-    val yesSelector = "#value"
-    val continueButtonSelector = "#continue"
-    val detailsTitleSelector = "#main-content > div > div > form > details > summary > span"
-    val detailsText1Selector = "#main-content > div > div > form > details > div > p:nth-child(1)"
-    val detailsText2Selector = "#main-content > div > div > form > details > div > p:nth-child(2)"
-    val detailsText3Selector = "#main-content > div > div > form > details > div > p:nth-child(3)"
-    val formSelector = "#main-content > div > div > form"
-  }
-
-  trait CommonExpectedResults {
-    val expectedCaption: Int => String
-    val expectedParagraphText: String
-    val yesText: String
-    val noText: String
-    val buttonText: String
-    val expectedDetailsTitle: String
-    val expectedDetailsText1: String
-    val expectedDetailsText3: String
-  }
-
-  trait SpecificExpectedResults {
-    val expectedTitle: String
-    val expectedHeading: String
-    val expectedErrorTitle: String
-    val expectedErrorMessage: String
-    val expectedDetailsText2: String
-  }
-
-  object CommonExpectedEN extends CommonExpectedResults {
-    val expectedCaption: Int => String = (taxYear: Int) => s"Employment benefits for 6 April ${taxYear - 1} to 5 April $taxYear"
-    val expectedParagraphText: String = "Living accommodation is any accommodation that you can live in, whether you live there all " +
-      "the time or only occasionally. It includes houses, flats, houseboats, holiday homes and apartments."
-    val yesText = "Yes"
-    val noText = "No"
-    val buttonText = "Continue"
-    val expectedDetailsTitle = "More information about living accommodation"
-    val expectedDetailsText1: String = "Living accommodation doesn’t include hotel rooms or board and lodgings, where you’re " +
-      "dependent on someone else for cooking, cleaning or laundry."
-    val expectedDetailsText3: String = "If you think all or part of this amount should be exempt from tax, refer to HS202 Living " +
-      "accommodation and follow the working sheet."
-  }
-
-  object CommonExpectedCY extends CommonExpectedResults {
-    val expectedCaption: Int => String = (taxYear: Int) => s"Employment benefits for 6 April ${taxYear - 1} to 5 April $taxYear"
-    val expectedParagraphText: String = "Diffinnir llety fel llety y gallwch fyw ynddo, píun a ydych yn byw ynddo drwyír amser neu dim ond rhan oír amser. " +
-      "Maeín cynnwys tai, fflatiau, cychod preswyl a lletyau gwyliau."
-    val yesText = "Iawn"
-    val noText = "Na"
-    val buttonText = "Yn eich blaen"
-    val expectedDetailsTitle = "Rhagor o wybodaeth am lety byw"
-    val expectedDetailsText1: String = "Nid yw llety yn cynnwys ystafelloedd mewn gwesty neu ëboard and lodgingsí, pan eich bod yn dibynnu ar rywun arall am goginio, glanhau neu olchi dillad."
-    val expectedDetailsText3: String = "Os ydych yn credu y dylaiír swm cyfan, neu ran ohono, fod wedi ei eithrio rhag treth, cyfeiriwch at HS202 Living accommodation a dilynwch y daflen waith."
-  }
-
-  object ExpectedIndividualEN extends SpecificExpectedResults {
-    val expectedTitle = "Did you get any living accommodation benefits?"
-    val expectedHeading = "Did you get any living accommodation benefits?"
-    val expectedErrorTitle = s"Error: $expectedTitle"
-    val expectedErrorMessage = "Select yes if you got living accommodation benefits"
-    val expectedDetailsText2: String = "Your employment income should include the value of any living accommodation you or your " +
-      "relations get because of your employment."
-  }
-
-  object ExpectedIndividualCY extends SpecificExpectedResults {
-    val expectedTitle = "A gawsoch unrhyw fuddiannau llety byw?"
-    val expectedHeading = "A gawsoch unrhyw fuddiannau llety byw?"
-    val expectedErrorTitle = s"Gwall: $expectedTitle"
-    val expectedErrorMessage = "Dewiswch ëIawní os cawsoch unrhyw fuddiannau llety byw"
-    val expectedDetailsText2: String = "Dylaiích incwm o gyflogaeth gynnwys gwerth unrhyw lety byw rydych chi neu eich perthnasau yn ei gael oherwydd eich cyflogaeth."
-  }
-
-  object ExpectedAgentEN extends SpecificExpectedResults {
-    val expectedTitle = "Did your client get any living accommodation benefits?"
-    val expectedHeading = "Did your client get any living accommodation benefits?"
-    val expectedErrorTitle = s"Error: $expectedTitle"
-    val expectedErrorMessage = "Select yes if your client got living accommodation benefits"
-    val expectedDetailsText2: String = "Your client’s employment income should include the value of any living accommodation they " +
-      "or their relations get because of their employment."
-  }
-
-  object ExpectedAgentCY extends SpecificExpectedResults {
-    val expectedTitle = "A gafodd eich cleient unrhyw fuddiannau llety byw?"
-    val expectedHeading = "A gafodd eich cleient unrhyw fuddiannau llety byw?"
-    val expectedErrorTitle = s"Gwall: $expectedTitle"
-    val expectedErrorMessage = "Dewiswch ëIawní os cafodd eich cleient fuddiannau llety byw"
-    val expectedDetailsText2: String = "Dylai incwm o gyflogaeth eich cleient gynnwys gwerth unrhyw lety y mae ef neu ei berthnasau yn ei gael oherwydd ei gyflogaeth."
-  }
-
-  val userScenarios: Seq[UserScenario[CommonExpectedResults, SpecificExpectedResults]] = {
-    Seq(UserScenario(isWelsh = false, isAgent = false, CommonExpectedEN, Some(ExpectedIndividualEN)),
-      UserScenario(isWelsh = false, isAgent = true, CommonExpectedEN, Some(ExpectedAgentEN)),
-      UserScenario(isWelsh = true, isAgent = false, CommonExpectedCY, Some(ExpectedIndividualCY)),
-      UserScenario(isWelsh = true, isAgent = true, CommonExpectedCY, Some(ExpectedAgentCY)))
-  }
+  override val userScenarios: Seq[UserScenario[_, _]] = Seq.empty
 
   ".show" should {
-    userScenarios.foreach { user =>
-      import Selectors._
-      import user.commonExpectedResults._
-      s"language is ${welshTest(user.isWelsh)} and request is from an ${agentTest(user.isAgent)}" should {
-        "render the 'Did you get any accommodation benefits?' page with no pre-filled radio buttons" which {
-          implicit lazy val result: WSResponse = {
-            authoriseAgentOrIndividual(user.isAgent)
-            dropEmploymentDB()
-            userDataStub(anIncomeTaxUserData, nino, taxYearEOY)
-            val benefitsViewModel = aBenefitsViewModel.copy(accommodationRelocationModel = Some(anAccommodationRelocationModel.copy(accommodationQuestion = None)))
-            insertCyaData(anEmploymentUserDataWithBenefits(benefitsViewModel))
-            urlGet(fullUrl(livingAccommodationBenefitsUrl(taxYearEOY, employmentId)), welsh = user.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
-          }
+    "render the 'Did you get any accommodation benefits?' page with no pre-filled radio buttons" which {
+      implicit lazy val result: WSResponse = {
+        authoriseAgentOrIndividual(isAgent = false)
+        dropEmploymentDB()
+        userDataStub(anIncomeTaxUserData, nino, taxYearEOY)
+        val benefitsViewModel = aBenefitsViewModel.copy(accommodationRelocationModel = Some(anAccommodationRelocationModel.copy(accommodationQuestion = None)))
+        insertCyaData(anEmploymentUserDataWithBenefits(benefitsViewModel))
+        urlGet(fullUrl(livingAccommodationBenefitsUrl(taxYearEOY, employmentId)), headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
+      }
 
-          s"has an OK($OK) status" in {
-            result.status shouldBe OK
-          }
-
-          lazy val document = Jsoup.parse(result.body)
-
-          implicit def documentSupplier: () => Document = () => document
-
-          titleCheck(user.specificExpectedResults.get.expectedTitle, user.isWelsh)
-          h1Check(user.specificExpectedResults.get.expectedHeading)
-          captionCheck(expectedCaption(taxYearEOY))
-          textOnPageCheck(expectedParagraphText, paragraphSelector)
-          radioButtonCheck(yesText, 1, checked = false)
-          radioButtonCheck(noText, 2, checked = false)
-          buttonCheck(buttonText, continueButtonSelector)
-          formPostLinkCheck(livingAccommodationBenefitsUrl(taxYearEOY, employmentId), formSelector)
-          textOnPageCheck(expectedDetailsTitle, detailsTitleSelector)
-          textOnPageCheck(expectedDetailsText1, detailsText1Selector)
-          textOnPageCheck(user.specificExpectedResults.get.expectedDetailsText2, detailsText2Selector)
-          textOnPageCheck(expectedDetailsText3, detailsText3Selector)
-          welshToggleCheck(user.isWelsh)
-        }
-
-        "render the 'Did you get any accommodation benefits?' page with the 'yes' radio button pre-filled and cya data" which {
-          implicit lazy val result: WSResponse = {
-            authoriseAgentOrIndividual(user.isAgent)
-            dropEmploymentDB()
-            userDataStub(anIncomeTaxUserData, nino, taxYearEOY)
-            val benefitsViewModel = aBenefitsViewModel.copy(accommodationRelocationModel = Some(anAccommodationRelocationModel.copy(accommodationQuestion = Some(true))))
-            insertCyaData(anEmploymentUserDataWithBenefits(benefitsViewModel))
-            urlGet(fullUrl(livingAccommodationBenefitsUrl(taxYearEOY, employmentId)), welsh = user.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
-          }
-
-          s"has an OK($OK) status" in {
-            result.status shouldBe OK
-          }
-
-          lazy val document = Jsoup.parse(result.body)
-
-          implicit def documentSupplier: () => Document = () => document
-
-          titleCheck(user.specificExpectedResults.get.expectedTitle, user.isWelsh)
-          h1Check(user.specificExpectedResults.get.expectedHeading)
-          captionCheck(expectedCaption(taxYearEOY))
-          textOnPageCheck(expectedParagraphText, paragraphSelector)
-          radioButtonCheck(yesText, 1, checked = true)
-          radioButtonCheck(noText, 2, checked = false)
-          buttonCheck(buttonText, continueButtonSelector)
-          formPostLinkCheck(livingAccommodationBenefitsUrl(taxYearEOY, employmentId), formSelector)
-          textOnPageCheck(expectedDetailsTitle, detailsTitleSelector)
-          textOnPageCheck(expectedDetailsText1, detailsText1Selector)
-          textOnPageCheck(user.specificExpectedResults.get.expectedDetailsText2, detailsText2Selector)
-          textOnPageCheck(expectedDetailsText3, detailsText3Selector)
-          welshToggleCheck(user.isWelsh)
-        }
-
-        "render the 'Did you get any accommodation benefits?' page with the 'no' radio button pre-filled and no prior benefits exist" which {
-          implicit lazy val result: WSResponse = {
-            authoriseAgentOrIndividual(user.isAgent)
-            dropEmploymentDB()
-            userDataStub(anIncomeTaxUserData, nino, taxYearEOY)
-            val benefitsViewModel = aBenefitsViewModel.copy(accommodationRelocationModel = Some(anAccommodationRelocationModel.copy(accommodationQuestion = Some(false))))
-            insertCyaData(anEmploymentUserDataWithBenefits(benefitsViewModel, isPriorSubmission = false, hasPriorBenefits = false))
-            urlGet(fullUrl(livingAccommodationBenefitsUrl(taxYearEOY, employmentId)), welsh = user.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
-          }
-
-          s"has an OK($OK) status" in {
-            result.status shouldBe OK
-          }
-
-          lazy val document = Jsoup.parse(result.body)
-
-          implicit def documentSupplier: () => Document = () => document
-
-          titleCheck(user.specificExpectedResults.get.expectedTitle, user.isWelsh)
-          h1Check(user.specificExpectedResults.get.expectedHeading)
-          captionCheck(expectedCaption(taxYearEOY))
-          textOnPageCheck(expectedParagraphText, paragraphSelector)
-          radioButtonCheck(yesText, 1, checked = false)
-          radioButtonCheck(noText, 2, checked = true)
-          buttonCheck(buttonText, continueButtonSelector)
-          formPostLinkCheck(livingAccommodationBenefitsUrl(taxYearEOY, employmentId), formSelector)
-          textOnPageCheck(expectedDetailsTitle, detailsTitleSelector)
-          textOnPageCheck(expectedDetailsText1, detailsText1Selector)
-          textOnPageCheck(user.specificExpectedResults.get.expectedDetailsText2, detailsText2Selector)
-          textOnPageCheck(expectedDetailsText3, detailsText3Selector)
-          welshToggleCheck(user.isWelsh)
-        }
+      s"has an OK($OK) status" in {
+        result.status shouldBe OK
       }
     }
 
@@ -304,37 +116,18 @@ class LivingAccommodationBenefitsControllerISpec extends IntegrationTest with Vi
   }
 
   ".submit" should {
-    userScenarios.foreach { user =>
-      import Selectors._
-      import user.commonExpectedResults._
+    "return an error when a form is submitted with no entry" which {
+      implicit lazy val result: WSResponse = {
+        authoriseAgentOrIndividual(isAgent = false)
+        dropEmploymentDB()
+        userDataStub(anIncomeTaxUserData, nino, taxYearEOY)
+        val benefitsViewModel = aBenefitsViewModel.copy(accommodationRelocationModel = Some(anAccommodationRelocationModel))
+        insertCyaData(anEmploymentUserDataWithBenefits(benefitsViewModel))
+        urlPost(fullUrl(livingAccommodationBenefitsUrl(taxYearEOY, employmentId)), body = "", headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
+      }
 
-      s"language is ${welshTest(user.isWelsh)} and request is from an ${agentTest(user.isAgent)}" should {
-        "return an error when a form is submitted with no entry" which {
-          implicit lazy val result: WSResponse = {
-            authoriseAgentOrIndividual(user.isAgent)
-            dropEmploymentDB()
-            userDataStub(anIncomeTaxUserData, nino, taxYearEOY)
-            val benefitsViewModel = aBenefitsViewModel.copy(accommodationRelocationModel = Some(anAccommodationRelocationModel))
-            insertCyaData(anEmploymentUserDataWithBenefits(benefitsViewModel))
-            urlPost(fullUrl(livingAccommodationBenefitsUrl(taxYearEOY, employmentId)), body = "", welsh = user.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
-          }
-
-          s"has an BAD REQUEST($BAD_REQUEST) status" in {
-            result.status shouldBe BAD_REQUEST
-          }
-
-          lazy val document = Jsoup.parse(result.body)
-
-          implicit def documentSupplier: () => Document = () => document
-
-          titleCheck(user.specificExpectedResults.get.expectedErrorTitle, user.isWelsh)
-          h1Check(user.specificExpectedResults.get.expectedHeading)
-          captionCheck(expectedCaption(taxYearEOY))
-          errorSummaryCheck(user.specificExpectedResults.get.expectedErrorMessage, yesSelector)
-          errorAboveElementCheck(user.specificExpectedResults.get.expectedErrorMessage, Some("value"))
-          formPostLinkCheck(livingAccommodationBenefitsUrl(taxYearEOY, employmentId), formSelector)
-          welshToggleCheck(user.isWelsh)
-        }
+      s"has an BAD REQUEST($BAD_REQUEST) status" in {
+        result.status shouldBe BAD_REQUEST
       }
     }
