+++ it/controllers/benefits/assets/AssetsBenefitsAmountControllerISpec.scala
@@ -18,8 +18,6 @@ package controllers.benefits.assets
 
 import forms.AmountForm
 import models.benefits.AssetsModel
-import org.jsoup.Jsoup
-import org.jsoup.nodes.Document
 import play.api.http.HeaderNames
 import play.api.http.Status.{BAD_REQUEST, OK, SEE_OTHER}
 import play.api.libs.ws.WSResponse
@@ -34,214 +32,23 @@ import utils.{EmploymentDatabaseHelper, IntegrationTest, ViewHelpers}
 class AssetsBenefitsAmountControllerISpec extends IntegrationTest with ViewHelpers with EmploymentDatabaseHelper {
 
   private val employmentId: String = "employmentId"
-  private val amount: BigDecimal = 100
-  private val amountInputName = "amount"
-  private val expectedErrorHref = "#amount"
   private val assetsSoFar: AssetsModel = AssetsModel(sectionQuestion = Some(true), assetsQuestion = Some(true), None, None, None)
 
-  object Selectors {
-    val hintTextSelector = "#amount-hint"
-    val currencyPrefixSelector = "#main-content > div > div > form > div > div.govuk-input__wrapper > div"
-    val inputSelector = "#amount"
-    val continueButtonSelector = "#continue"
-    val formSelector = "#main-content > div > div > form"
-    val youCanSelector = "#you-can-text"
-    val enterTotalSelector = "#enter-total-text"
-    val previousAmountSelector = "#previous-amount-text"
-  }
-
-  trait CommonExpectedResults {
-    val expectedCaption: String
-
-    def optionalParagraphText(amount: BigDecimal): String
-
-    val expectedHintText: String
-    val currencyPrefix: String
-    val continueButtonText: String
-    val enterTotalText: String
-  }
-
-  trait SpecificExpectedResults {
-    val expectedTitle: String
-    val expectedHeading: String
-    val expectedYouCanText: String
-    val expectedErrorTitle: String
-    val expectedErrorNoEntry: String
-    val expectedErrorIncorrectFormat: String
-    val expectedErrorOverMaximum: String
-  }
-
-  object CommonExpectedEN extends CommonExpectedResults {
-    val expectedCaption: String = s"Employment benefits for 6 April ${taxYearEOY - 1} to 5 April $taxYearEOY"
-
-    def optionalParagraphText(amount: BigDecimal): String = s"If it was not £$amount, tell us the correct amount."
-
-    val expectedHintText = "For example, £193.52"
-    val currencyPrefix = "£"
-    val continueButtonText = "Continue"
-    val enterTotalText = "Enter the total."
-  }
-
-  object CommonExpectedCY extends CommonExpectedResults {
-    val expectedCaption: String = s"Employment benefits for 6 April ${taxYearEOY - 1} to 5 April $taxYearEOY"
-
-    def optionalParagraphText(amount: BigDecimal): String = s"Rhowch wybod y swm cywir os nad oedd yn £$amount."
-
-    val expectedHintText = "Er enghraifft, £193.52"
-    val currencyPrefix = "£"
-    val continueButtonText = "Yn eich blaen"
-    val enterTotalText = "Nodwch y cyfanswm."
-  }
-
-  object ExpectedIndividualEN extends SpecificExpectedResults {
-    val expectedTitle = "How much were the assets made available for your use?"
-    val expectedHeading = "How much were the assets made available for your use?"
-    val expectedErrorTitle = s"Error: $expectedTitle"
-    val expectedYouCanText: String = "You can find this information on your P11D form in section L, box 13."
-    val expectedErrorNoEntry = "Enter the amount for assets made available for your use"
-    val expectedErrorIncorrectFormat = "Enter the amount for assets made available for your use in the correct format"
-    val expectedErrorOverMaximum = "The amount for assets made available for your use must be less than £100,000,000,000"
-  }
-
-  object ExpectedIndividualCY extends SpecificExpectedResults {
-    val expectedTitle = "Beth oedd gwerth yr asedion a oedd ar gael i chi eu defnyddio?"
-    val expectedHeading = "Beth oedd gwerth yr asedion a oedd ar gael i chi eu defnyddio?"
-    val expectedErrorTitle = s"Gwall: $expectedTitle"
-    val expectedYouCanText: String = "Maeír wybodaeth hon ar gael yn adran L, blwch 13 ar eich ffurflen P11D."
-    val expectedErrorNoEntry = "Nodwch swm yr asedion a oedd ar gael i chi eu defnyddio"
-    val expectedErrorIncorrectFormat = "Nodwch swm yr asedion a oedd ar gael i chi eu defnyddio yn y fformat cywir"
-    val expectedErrorOverMaximum = "Maeín rhaid i swm yr asedion a oedd ar gael i chi eu defnyddio fod yn llai na £100,000,000,000"
-  }
-
-  object ExpectedAgentEN extends SpecificExpectedResults {
-    val expectedTitle = "How much were the assets made available for your client’s use?"
-    val expectedHeading = "How much were the assets made available for your client’s use?"
-    val expectedErrorTitle = s"Error: $expectedTitle"
-    val expectedYouCanText: String = "You can find this information on your client’s P11D form in section L, box 13."
-    val expectedErrorNoEntry = "Enter the amount for assets made available for your client’s use"
-    val expectedErrorIncorrectFormat = "Enter the amount for assets made available for your client’s use in the correct format"
-    val expectedErrorOverMaximum = "The amount for assets made available for your client’s use must be less than £100,000,000,000"
-  }
-
-  object ExpectedAgentCY extends SpecificExpectedResults {
-    val expectedTitle = "Beth oedd gwerth yr asedion a oedd ar gael iích cleient eu defnyddio?"
-    val expectedHeading = "Beth oedd gwerth yr asedion a oedd ar gael iích cleient eu defnyddio?"
-    val expectedErrorTitle = s"Gwall: $expectedTitle"
-    val expectedYouCanText: String = "Maeír wybodaeth hon ar gael yn adran L, blwch 13 ar ffurflen P11D eich cleient."
-    val expectedErrorNoEntry = "Nodwch swm yr asedion a oedd ar gael iích cleient eu defnyddio"
-    val expectedErrorIncorrectFormat = "Nodwch swm yr asedion a oedd ar gael iích cleient eu defnyddio yn y fformat cywir"
-    val expectedErrorOverMaximum = "Maeín rhaid i swm yr asedion a oedd ar gael iích cleient eu defnyddio fod yn llai na £100,000,000,000"
-  }
-
-  val userScenarios: Seq[UserScenario[CommonExpectedResults, SpecificExpectedResults]] = {
-    Seq(UserScenario(isWelsh = false, isAgent = false, CommonExpectedEN, Some(ExpectedIndividualEN)),
-      UserScenario(isWelsh = false, isAgent = true, CommonExpectedEN, Some(ExpectedAgentEN)),
-      UserScenario(isWelsh = true, isAgent = false, CommonExpectedCY, Some(ExpectedIndividualCY)),
-      UserScenario(isWelsh = true, isAgent = true, CommonExpectedCY, Some(ExpectedAgentCY))
-    )
-  }
+  override val userScenarios: Seq[UserScenario[_, _]] = Seq.empty
 
   ".show" should {
-    userScenarios.foreach { user =>
-      import Selectors._
-      import user.commonExpectedResults._
+    "render the assets amount page with no prefilled data" which {
+      implicit lazy val result: WSResponse = {
+        authoriseAgentOrIndividual(isAgent = false)
+        dropEmploymentDB()
+        userDataStub(anIncomeTaxUserData, nino, taxYearEOY)
+        val benefitsViewModel = aBenefitsViewModel.copy(assetsModel = Some(anAssetsModel.copy(assets = None)))
+        insertCyaData(anEmploymentUserDataWithBenefits(benefitsViewModel))
+        urlGet(fullUrl(assetsForUseBenefitsAmountUrl(taxYearEOY, employmentId)), headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
+      }
 
-      s"language is ${welshTest(user.isWelsh)} and request is from an ${agentTest(user.isAgent)}" should {
-        "render the assets amount page with no prefilled data" which {
-          implicit lazy val result: WSResponse = {
-            authoriseAgentOrIndividual(user.isAgent)
-            dropEmploymentDB()
-            userDataStub(anIncomeTaxUserData, nino, taxYearEOY)
-            val benefitsViewModel = aBenefitsViewModel.copy(assetsModel = Some(anAssetsModel.copy(assets = None)))
-            insertCyaData(anEmploymentUserDataWithBenefits(benefitsViewModel))
-            urlGet(fullUrl(assetsForUseBenefitsAmountUrl(taxYearEOY, employmentId)), welsh = user.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
-          }
-
-          s"has an OK($OK) status" in {
-            result.status shouldBe OK
-          }
-
-          lazy val document = Jsoup.parse(result.body)
-
-          implicit def documentSupplier: () => Document = () => document
-
-          titleCheck(user.specificExpectedResults.get.expectedTitle, user.isWelsh)
-          h1Check(user.specificExpectedResults.get.expectedHeading)
-          captionCheck(expectedCaption)
-          elementsNotOnPageCheck(previousAmountSelector)
-          textOnPageCheck(enterTotalText, enterTotalSelector)
-          textOnPageCheck(user.specificExpectedResults.get.expectedYouCanText, youCanSelector)
-          textOnPageCheck(expectedHintText, hintTextSelector)
-          textOnPageCheck(currencyPrefix, currencyPrefixSelector)
-          inputFieldValueCheck(amountInputName, inputSelector, "")
-          buttonCheck(continueButtonText, continueButtonSelector)
-          formPostLinkCheck(assetsForUseBenefitsAmountUrl(taxYearEOY, employmentId), formSelector)
-
-          welshToggleCheck(user.isWelsh)
-        }
-
-        "render the assets amount page with a pre-filled amount field when there's cya data and prior benefits" which {
-          implicit lazy val result: WSResponse = {
-            authoriseAgentOrIndividual(user.isAgent)
-            dropEmploymentDB()
-            userDataStub(anIncomeTaxUserData, nino, taxYearEOY)
-            insertCyaData(anEmploymentUserData)
-            urlGet(fullUrl(assetsForUseBenefitsAmountUrl(taxYearEOY, employmentId)), welsh = user.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
-          }
-
-          s"has an OK($OK) status" in {
-            result.status shouldBe OK
-          }
-
-          lazy val document = Jsoup.parse(result.body)
-
-          implicit def documentSupplier: () => Document = () => document
-
-          titleCheck(user.specificExpectedResults.get.expectedTitle, user.isWelsh)
-          h1Check(user.specificExpectedResults.get.expectedHeading)
-          captionCheck(expectedCaption)
-          textOnPageCheck(optionalParagraphText(amount), previousAmountSelector)
-          textOnPageCheck(enterTotalText, enterTotalSelector)
-          textOnPageCheck(user.specificExpectedResults.get.expectedYouCanText, youCanSelector)
-          textOnPageCheck(expectedHintText, hintTextSelector)
-          textOnPageCheck(currencyPrefix, currencyPrefixSelector)
-          inputFieldValueCheck(amountInputName, inputSelector, amount.toString())
-          buttonCheck(continueButtonText, continueButtonSelector)
-          formPostLinkCheck(assetsForUseBenefitsAmountUrl(taxYearEOY, employmentId), formSelector)
-
-          welshToggleCheck(user.isWelsh)
-        }
-
-        "render the assets amount page with a pre-filled amount field when there's cya data and no prior benefits" which {
-          implicit lazy val result: WSResponse = {
-            authoriseAgentOrIndividual(user.isAgent)
-            dropEmploymentDB()
-            insertCyaData(anEmploymentUserData.copy(isPriorSubmission = false, hasPriorBenefits = false))
-            urlGet(fullUrl(assetsForUseBenefitsAmountUrl(taxYearEOY, employmentId)), welsh = user.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
-          }
-
-          s"has an OK($OK) status" in {
-            result.status shouldBe OK
-          }
-
-          lazy val document = Jsoup.parse(result.body)
-
-          implicit def documentSupplier: () => Document = () => document
-
-          titleCheck(user.specificExpectedResults.get.expectedTitle, user.isWelsh)
-          h1Check(user.specificExpectedResults.get.expectedHeading)
-          captionCheck(expectedCaption)
-          textOnPageCheck(optionalParagraphText(amount), previousAmountSelector)
-          textOnPageCheck(enterTotalText, enterTotalSelector)
-          textOnPageCheck(user.specificExpectedResults.get.expectedYouCanText, youCanSelector)
-          textOnPageCheck(expectedHintText, hintTextSelector)
-          textOnPageCheck(currencyPrefix, currencyPrefixSelector)
-          inputFieldValueCheck(amountInputName, inputSelector, amount.toString())
-          buttonCheck(continueButtonText, continueButtonSelector)
-          formPostLinkCheck(assetsForUseBenefitsAmountUrl(taxYearEOY, employmentId), formSelector)
-
-          welshToggleCheck(user.isWelsh)
-        }
+      s"has an OK($OK) status" in {
+        result.status shouldBe OK
       }
     }
 
@@ -323,122 +130,18 @@ class AssetsBenefitsAmountControllerISpec extends IntegrationTest with ViewHelpe
   }
 
   ".submit" should {
-    userScenarios.foreach { user =>
-      import Selectors._
-      import user.commonExpectedResults._
+    "render the assets amount page with an error" when {
+      "a form is submitted with no entry" which {
+        implicit lazy val result: WSResponse = {
+          authoriseAgentOrIndividual(isAgent = false)
+          dropEmploymentDB()
+          userDataStub(anIncomeTaxUserData, nino, taxYearEOY)
+          insertCyaData(anEmploymentUserData)
+          urlPost(fullUrl(assetsForUseBenefitsAmountUrl(taxYearEOY, employmentId)), body = "", headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
+        }
 
-      s"language is ${welshTest(user.isWelsh)} and request is from an ${agentTest(user.isAgent)}" should {
-        "render the assets amount page with an error" when {
-          "a form is submitted with no entry" which {
-            implicit lazy val result: WSResponse = {
-              authoriseAgentOrIndividual(user.isAgent)
-              dropEmploymentDB()
-              userDataStub(anIncomeTaxUserData, nino, taxYearEOY)
-              insertCyaData(anEmploymentUserData)
-              urlPost(fullUrl(assetsForUseBenefitsAmountUrl(taxYearEOY, employmentId)), body = "", welsh = user.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
-            }
-
-            s"has an BAD REQUEST($BAD_REQUEST) status" in {
-              result.status shouldBe BAD_REQUEST
-            }
-
-            lazy val document = Jsoup.parse(result.body)
-
-            implicit def documentSupplier: () => Document = () => document
-
-            titleCheck(user.specificExpectedResults.get.expectedErrorTitle, user.isWelsh)
-            h1Check(user.specificExpectedResults.get.expectedHeading)
-            captionCheck(expectedCaption)
-            textOnPageCheck(optionalParagraphText(amount), previousAmountSelector)
-            textOnPageCheck(enterTotalText, enterTotalSelector)
-            textOnPageCheck(user.specificExpectedResults.get.expectedYouCanText, youCanSelector)
-            textOnPageCheck(expectedHintText, hintTextSelector)
-            textOnPageCheck(currencyPrefix, currencyPrefixSelector)
-            inputFieldValueCheck(amountInputName, inputSelector, "")
-            buttonCheck(continueButtonText, continueButtonSelector)
-            formPostLinkCheck(assetsForUseBenefitsAmountUrl(taxYearEOY, employmentId), formSelector)
-
-            errorSummaryCheck(user.specificExpectedResults.get.expectedErrorNoEntry, expectedErrorHref)
-            errorAboveElementCheck(user.specificExpectedResults.get.expectedErrorNoEntry, Some(amountInputName))
-
-            welshToggleCheck(user.isWelsh)
-          }
-
-          "a form is submitted with an incorrectly formatted amount" which {
-            val incorrectFormatAmount = "abc"
-            val form: Map[String, String] = Map(AmountForm.amount -> incorrectFormatAmount)
-
-            implicit lazy val result: WSResponse = {
-              authoriseAgentOrIndividual(user.isAgent)
-              dropEmploymentDB()
-              userDataStub(anIncomeTaxUserData, nino, taxYearEOY)
-              insertCyaData(anEmploymentUserData)
-              urlPost(fullUrl(assetsForUseBenefitsAmountUrl(taxYearEOY, employmentId)), body = form, welsh = user.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
-            }
-
-            s"has an BAD REQUEST($BAD_REQUEST) status" in {
-              result.status shouldBe BAD_REQUEST
-            }
-
-            lazy val document = Jsoup.parse(result.body)
-
-            implicit def documentSupplier: () => Document = () => document
-
-            titleCheck(user.specificExpectedResults.get.expectedErrorTitle, user.isWelsh)
-            h1Check(user.specificExpectedResults.get.expectedHeading)
-            captionCheck(expectedCaption)
-            textOnPageCheck(optionalParagraphText(amount), previousAmountSelector)
-            textOnPageCheck(enterTotalText, enterTotalSelector)
-            textOnPageCheck(user.specificExpectedResults.get.expectedYouCanText, youCanSelector)
-            textOnPageCheck(expectedHintText, hintTextSelector)
-            textOnPageCheck(currencyPrefix, currencyPrefixSelector)
-            inputFieldValueCheck(amountInputName, inputSelector, incorrectFormatAmount)
-            buttonCheck(continueButtonText, continueButtonSelector)
-            formPostLinkCheck(assetsForUseBenefitsAmountUrl(taxYearEOY, employmentId), formSelector)
-
-            errorSummaryCheck(user.specificExpectedResults.get.expectedErrorIncorrectFormat, expectedErrorHref)
-            errorAboveElementCheck(user.specificExpectedResults.get.expectedErrorIncorrectFormat, Some(amountInputName))
-
-            welshToggleCheck(user.isWelsh)
-          }
-
-          "a form is submitted and the amount is over the maximum limit" which {
-            val overMaximumAmount = "100,000,000,000,000,000,000"
-            val form: Map[String, String] = Map(AmountForm.amount -> overMaximumAmount)
-
-            implicit lazy val result: WSResponse = {
-              authoriseAgentOrIndividual(user.isAgent)
-              dropEmploymentDB()
-              userDataStub(anIncomeTaxUserData, nino, taxYearEOY)
-              insertCyaData(anEmploymentUserData)
-              urlPost(fullUrl(assetsForUseBenefitsAmountUrl(taxYearEOY, employmentId)), body = form, welsh = user.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
-            }
-
-            s"has an BAD REQUEST($BAD_REQUEST) status" in {
-              result.status shouldBe BAD_REQUEST
-            }
-
-            lazy val document = Jsoup.parse(result.body)
-
-            implicit def documentSupplier: () => Document = () => document
-
-            titleCheck(user.specificExpectedResults.get.expectedErrorTitle, user.isWelsh)
-            h1Check(user.specificExpectedResults.get.expectedHeading)
-            captionCheck(expectedCaption)
-            textOnPageCheck(optionalParagraphText(amount), previousAmountSelector)
-            textOnPageCheck(enterTotalText, enterTotalSelector)
-            textOnPageCheck(user.specificExpectedResults.get.expectedYouCanText, youCanSelector)
-            textOnPageCheck(expectedHintText, hintTextSelector)
-            textOnPageCheck(currencyPrefix, currencyPrefixSelector)
-            inputFieldValueCheck(amountInputName, inputSelector, overMaximumAmount)
-            buttonCheck(continueButtonText, continueButtonSelector)
-            formPostLinkCheck(assetsForUseBenefitsAmountUrl(taxYearEOY, employmentId), formSelector)
-
-            errorSummaryCheck(user.specificExpectedResults.get.expectedErrorOverMaximum, expectedErrorHref)
-            errorAboveElementCheck(user.specificExpectedResults.get.expectedErrorOverMaximum, Some(amountInputName))
-
-            welshToggleCheck(user.isWelsh)
-          }
+        s"has an BAD REQUEST($BAD_REQUEST) status" in {
+          result.status shouldBe BAD_REQUEST
         }
       }
     }
