+++ test/views/benefits/assets/AssetTransfersBenefitsViewSpec.scala
@@ -1,0 +1,179 @@
+/*
+ * Copyright 2022 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package views.benefits.assets
+
+import controllers.benefits.assets.routes.AssetTransfersBenefitsController
+import forms.YesNoForm
+import forms.benefits.assets.AssetsFormsProvider
+import models.AuthorisationRequest
+import org.jsoup.Jsoup
+import org.jsoup.nodes.Document
+import play.api.data.Form
+import play.api.i18n.Messages
+import play.api.mvc.AnyContent
+import support.ViewUnitTest
+import views.html.benefits.assets.AssetTransfersBenefitsView
+
+class AssetTransfersBenefitsViewSpec extends ViewUnitTest {
+
+  private val employmentId: String = "employmentId"
+
+  object Selectors {
+    val paragraphSelector: String = "#main-content > div > div > p"
+    val continueButtonSelector: String = "#continue"
+    val continueButtonFormSelector: String = "#main-content > div > div > form"
+    val yesSelector = "#value"
+  }
+
+  trait SpecificExpectedResults {
+    val expectedTitle: String
+    val expectedHeading: String
+    val expectedParagraph: String
+    val expectedErrorTitle: String
+    val expectedErrorText: String
+  }
+
+  trait CommonExpectedResults {
+    val expectedCaption: String
+    val expectedButtonText: String
+    val yesText: String
+    val noText: String
+  }
+
+  object ExpectedIndividualEN extends SpecificExpectedResults {
+    val expectedTitle = "Did your employer give you any assets to keep?"
+    val expectedHeading = "Did your employer give you any assets to keep?"
+    val expectedParagraph = "You became the owner of these assets."
+    val expectedErrorTitle = s"Error: $expectedTitle"
+    val expectedErrorText = "Select yes if your employer gave you assets to keep"
+  }
+
+  object ExpectedIndividualCY extends SpecificExpectedResults {
+    val expectedTitle = "A roddodd eich cyflogwr unrhyw asedion i chi eu cadw?"
+    val expectedHeading = "A roddodd eich cyflogwr unrhyw asedion i chi eu cadw?"
+    val expectedParagraph = "Daethoch yn berchennog yr asedion hyn."
+    val expectedErrorTitle = s"Gwall: $expectedTitle"
+    val expectedErrorText = "Dewiswch ëIawní os rhoddodd eich cyflogwr asedion i chi eu cadw"
+  }
+
+  object ExpectedAgentEN extends SpecificExpectedResults {
+    val expectedTitle = "Did your client’s employer give them any assets to keep?"
+    val expectedHeading = "Did your client’s employer give them any assets to keep?"
+    val expectedParagraph = "Your client became the owner of these assets."
+    val expectedErrorTitle = s"Error: $expectedTitle"
+    val expectedErrorText = "Select yes if your client’s employer gave them assets to keep"
+  }
+
+  object ExpectedAgentCY extends SpecificExpectedResults {
+    val expectedTitle = "A roddodd cyflogwr eich cleient unrhyw asedion iddo eu cadw?"
+    val expectedHeading = "A roddodd cyflogwr eich cleient unrhyw asedion iddo eu cadw?"
+    val expectedParagraph = "Daeth eich cleient yn berchennog yr asedion hyn."
+    val expectedErrorTitle = s"Gwall: $expectedTitle"
+    val expectedErrorText = "Dewiswch ëIawní os rhoddodd cyflogwr eich cleient asedion iddo eu cadw"
+  }
+
+  object CommonExpectedEN extends CommonExpectedResults {
+    val expectedCaption = s"Employment benefits for 6 April ${taxYearEOY - 1} to 5 April $taxYearEOY"
+    val expectedButtonText = "Continue"
+    val yesText = "Yes"
+    val noText = "No"
+  }
+
+  object CommonExpectedCY extends CommonExpectedResults {
+    val expectedCaption = s"Employment benefits for 6 April ${taxYearEOY - 1} to 5 April $taxYearEOY"
+    val expectedButtonText = "Yn eich blaen"
+    val yesText = "Iawn"
+    val noText = "Na"
+  }
+
+  val userScenarios: Seq[UserScenario[CommonExpectedResults, SpecificExpectedResults]] = Seq(
+    UserScenario(isWelsh = false, isAgent = false, CommonExpectedEN, Some(ExpectedIndividualEN)),
+    UserScenario(isWelsh = false, isAgent = true, CommonExpectedEN, Some(ExpectedAgentEN)),
+    UserScenario(isWelsh = true, isAgent = false, CommonExpectedCY, Some(ExpectedIndividualCY)),
+    UserScenario(isWelsh = true, isAgent = true, CommonExpectedCY, Some(ExpectedAgentCY))
+  )
+
+  private def form(isAgent: Boolean): Form[Boolean] = new AssetsFormsProvider().assetTransfersForm(isAgent)
+
+  private lazy val underTest = inject[AssetTransfersBenefitsView]
+
+  userScenarios.foreach { userScenario =>
+    import Selectors._
+    import userScenario.commonExpectedResults._
+    s"language is ${welshTest(userScenario.isWelsh)} and request is from an ${agentTest(userScenario.isAgent)}" should {
+      "render 'asset transfers' yes/no page with the correct content with no pre-filling" which {
+        implicit val authRequest: AuthorisationRequest[AnyContent] = getAuthRequest(userScenario.isAgent)
+        implicit val messages: Messages = getMessages(userScenario.isWelsh)
+
+        val htmlFormat = underTest(form(userScenario.isAgent), taxYearEOY, employmentId)
+
+        implicit val document: Document = Jsoup.parse(htmlFormat.body)
+
+        titleCheck(userScenario.specificExpectedResults.get.expectedTitle, userScenario.isWelsh)
+        h1Check(userScenario.specificExpectedResults.get.expectedHeading)
+        captionCheck(expectedCaption)
+        textOnPageCheck(userScenario.specificExpectedResults.get.expectedParagraph, paragraphSelector)
+        radioButtonCheck(yesText, radioNumber = 1, checked = false)
+        radioButtonCheck(noText, radioNumber = 2, checked = false)
+        buttonCheck(expectedButtonText, continueButtonSelector)
+        formPostLinkCheck(AssetTransfersBenefitsController.submit(taxYearEOY, employmentId).url, continueButtonFormSelector)
+        welshToggleCheck(userScenario.isWelsh)
+      }
+
+      "render 'asset transfers' yes/no page with the correct content with yes pre-filled" which {
+        implicit val authRequest: AuthorisationRequest[AnyContent] = getAuthRequest(userScenario.isAgent)
+        implicit val messages: Messages = getMessages(userScenario.isWelsh)
+
+        val htmlFormat = underTest(form(userScenario.isAgent).fill(value = true), taxYearEOY, employmentId)
+
+        implicit val document: Document = Jsoup.parse(htmlFormat.body)
+
+        titleCheck(userScenario.specificExpectedResults.get.expectedTitle, userScenario.isWelsh)
+        h1Check(userScenario.specificExpectedResults.get.expectedHeading)
+        captionCheck(expectedCaption)
+        textOnPageCheck(userScenario.specificExpectedResults.get.expectedParagraph, paragraphSelector)
+        radioButtonCheck(yesText, radioNumber = 1, checked = true)
+        radioButtonCheck(noText, radioNumber = 2, checked = false)
+        buttonCheck(expectedButtonText, continueButtonSelector)
+        formPostLinkCheck(AssetTransfersBenefitsController.submit(taxYearEOY, employmentId).url, continueButtonFormSelector)
+        welshToggleCheck(userScenario.isWelsh)
+      }
+
+      "return an error when a user submits an empty form" which {
+        implicit val authRequest: AuthorisationRequest[AnyContent] = getAuthRequest(userScenario.isAgent)
+        implicit val messages: Messages = getMessages(userScenario.isWelsh)
+
+        val htmlFormat = underTest(form(userScenario.isAgent).bind(Map(YesNoForm.yesNo -> "")), taxYearEOY, employmentId)
+
+        implicit val document: Document = Jsoup.parse(htmlFormat.body)
+
+        titleCheck(userScenario.specificExpectedResults.get.expectedErrorTitle, userScenario.isWelsh)
+        h1Check(userScenario.specificExpectedResults.get.expectedHeading)
+        captionCheck(expectedCaption)
+        textOnPageCheck(userScenario.specificExpectedResults.get.expectedParagraph, paragraphSelector)
+        radioButtonCheck(yesText, radioNumber = 1, checked = false)
+        radioButtonCheck(noText, radioNumber = 2, checked = false)
+        buttonCheck(expectedButtonText, continueButtonSelector)
+        formPostLinkCheck(AssetTransfersBenefitsController.submit(taxYearEOY, employmentId).url, continueButtonFormSelector)
+        welshToggleCheck(userScenario.isWelsh)
+
+        errorSummaryCheck(userScenario.specificExpectedResults.get.expectedErrorText, Selectors.yesSelector)
+        errorAboveElementCheck(userScenario.specificExpectedResults.get.expectedErrorText, Some("value"))
+      }
+    }
+  }
+}
