+++ test/views/benefits/assets/AssetsBenefitsViewSpec.scala
@@ -1,0 +1,200 @@
+/*
+ * Copyright 2022 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package views.benefits.assets
+
+import controllers.benefits.assets.routes.AssetsBenefitsController
+import forms.YesNoForm
+import forms.benefits.assets.AssetsFormsProvider
+import models.AuthorisationRequest
+import org.jsoup.Jsoup
+import org.jsoup.nodes.Document
+import play.api.data.Form
+import play.api.i18n.Messages
+import play.api.mvc.AnyContent
+import support.ViewUnitTest
+import views.html.benefits.assets.AssetsBenefitsView
+
+class AssetsBenefitsViewSpec extends ViewUnitTest {
+
+  private val employmentId: String = "employmentId"
+
+  object Selectors {
+    val youCanUseTextSelector = "#main-content > div > div > p"
+    val yesSelector = "#value"
+    val continueButtonSelector = "#continue"
+    val formSelector = "#main-content > div > div > form"
+  }
+
+  trait CommonExpectedResults {
+    val expectedCaption: String
+    val expectedButtonText: String
+    val yesText: String
+    val noText: String
+    val continueText: String
+  }
+
+  trait SpecificExpectedResults {
+    val expectedTitle: String
+    val expectedHeading: String
+    val youCanUseText: String
+    val expectedErrorTitle: String
+    val expectedErrorText: String
+  }
+
+  object CommonExpectedEN extends CommonExpectedResults {
+    val expectedCaption = s"Employment benefits for 6 April ${taxYearEOY - 1} to 5 April $taxYearEOY"
+    val expectedButtonText = "Continue"
+    val yesText = "Yes"
+    val noText = "No"
+    val continueText = "Continue"
+  }
+
+  object CommonExpectedCY extends CommonExpectedResults {
+    val expectedCaption = s"Employment benefits for 6 April ${taxYearEOY - 1} to 5 April $taxYearEOY"
+    val expectedButtonText = "Yn eich blaen"
+    val yesText = "Iawn"
+    val noText = "Na"
+    val continueText = "Yn eich blaen"
+  }
+
+  object ExpectedIndividualEN extends SpecificExpectedResults {
+    val expectedTitle = "Did your employer make any assets available for your use?"
+    val expectedHeading = "Did your employer make any assets available for your use?"
+    val youCanUseText = "You can use these assets but you do not own them."
+    val expectedErrorTitle = s"Error: $expectedTitle"
+    val expectedErrorText = "Select yes if your employer made assets available for your use"
+  }
+
+  object ExpectedIndividualCY extends SpecificExpectedResults {
+    val expectedTitle = "A oedd eich cyflogwr wedi gwneud asedion ar gael i chi eu defnyddio?"
+    val expectedHeading = "A oedd eich cyflogwr wedi gwneud asedion ar gael i chi eu defnyddio?"
+    val youCanUseText = "Gallwch ddefnyddioír asedion hyn ond nid ydych yn berchen arnynt."
+    val expectedErrorTitle = s"Gwall: $expectedTitle"
+    val expectedErrorText = "Dewiswch ëIawní os oedd eich cyflogwr wedi gwneud asedion ar gael i chi eu defnyddio"
+  }
+
+  object ExpectedAgentEN extends SpecificExpectedResults {
+    val expectedTitle = "Did your client’s employer make any assets available for their use?"
+    val expectedHeading = "Did your client’s employer make any assets available for their use?"
+    val youCanUseText = "They can use these assets but they do not own them."
+    val expectedErrorTitle = s"Error: $expectedTitle"
+    val expectedErrorText = "Select yes if your client’s employer made assets available for their use"
+  }
+
+  object ExpectedAgentCY extends SpecificExpectedResults {
+    val expectedTitle = "A oedd cyflogwr eich cleient wedi gwneud asedion ar gael iddo eu defnyddio?"
+    val expectedHeading = "A oedd cyflogwr eich cleient wedi gwneud asedion ar gael iddo eu defnyddio?"
+    val youCanUseText = "Gall ddefnyddioír asedion hyn ond nid ywín berchen arnynt."
+    val expectedErrorTitle = s"Gwall: $expectedTitle"
+    val expectedErrorText = "Dewiswch ëIawní os oedd cyflogwr eich cleient wedi gwneud asedion ar gael iddo eu defnyddio"
+  }
+
+  val userScenarios: Seq[UserScenario[CommonExpectedResults, SpecificExpectedResults]] = Seq(
+    UserScenario(isWelsh = false, isAgent = false, CommonExpectedEN, Some(ExpectedIndividualEN)),
+    UserScenario(isWelsh = false, isAgent = true, CommonExpectedEN, Some(ExpectedAgentEN)),
+    UserScenario(isWelsh = true, isAgent = false, CommonExpectedCY, Some(ExpectedIndividualCY)),
+    UserScenario(isWelsh = true, isAgent = true, CommonExpectedCY, Some(ExpectedAgentCY))
+  )
+
+  private def form(isAgent: Boolean): Form[Boolean] = new AssetsFormsProvider().assetsForm(isAgent)
+
+  private lazy val underTest = inject[AssetsBenefitsView]
+
+  userScenarios.foreach { userScenario =>
+    import Selectors._
+    import userScenario.commonExpectedResults._
+    s"language is ${welshTest(userScenario.isWelsh)} and request is from an ${agentTest(userScenario.isAgent)}" should {
+      "render page without pre-filled radio buttons and the user doesn't have prior benefits" which {
+        implicit val authRequest: AuthorisationRequest[AnyContent] = getAuthRequest(userScenario.isAgent)
+        implicit val messages: Messages = getMessages(userScenario.isWelsh)
+
+        val htmlFormat = underTest(form(userScenario.isAgent), taxYearEOY, employmentId)
+
+        implicit val document: Document = Jsoup.parse(htmlFormat.body)
+
+        titleCheck(userScenario.specificExpectedResults.get.expectedTitle, userScenario.isWelsh)
+        h1Check(userScenario.specificExpectedResults.get.expectedHeading)
+        captionCheck(expectedCaption)
+        textOnPageCheck(userScenario.specificExpectedResults.get.youCanUseText, youCanUseTextSelector)
+        radioButtonCheck(yesText, radioNumber = 1, checked = false)
+        radioButtonCheck(noText, radioNumber = 2, checked = false)
+        buttonCheck(continueText, continueButtonSelector)
+        formPostLinkCheck(AssetsBenefitsController.submit(taxYearEOY, employmentId).url, formSelector)
+        welshToggleCheck(userScenario.isWelsh)
+      }
+
+      "render the assets benefits page with the 'yes' radio button prefilled and the user has prior benefits" which {
+        implicit val authRequest: AuthorisationRequest[AnyContent] = getAuthRequest(userScenario.isAgent)
+        implicit val messages: Messages = getMessages(userScenario.isWelsh)
+
+        val htmlFormat = underTest(form(userScenario.isAgent).fill(value = true), taxYearEOY, employmentId)
+
+        implicit val document: Document = Jsoup.parse(htmlFormat.body)
+
+        titleCheck(userScenario.specificExpectedResults.get.expectedTitle, userScenario.isWelsh)
+        h1Check(userScenario.specificExpectedResults.get.expectedHeading)
+        captionCheck(expectedCaption)
+        textOnPageCheck(userScenario.specificExpectedResults.get.youCanUseText, youCanUseTextSelector)
+        radioButtonCheck(yesText, radioNumber = 1, checked = true)
+        radioButtonCheck(noText, radioNumber = 2, checked = false)
+        buttonCheck(continueText, continueButtonSelector)
+        formPostLinkCheck(AssetsBenefitsController.submit(taxYearEOY, employmentId).url, formSelector)
+        welshToggleCheck(userScenario.isWelsh)
+      }
+
+      "render the assets benefits page with the 'no' radio button prefilled and the user doesn't have prior benefits" which {
+        implicit val authRequest: AuthorisationRequest[AnyContent] = getAuthRequest(userScenario.isAgent)
+        implicit val messages: Messages = getMessages(userScenario.isWelsh)
+
+        val htmlFormat = underTest(form(userScenario.isAgent).fill(value = false), taxYearEOY, employmentId)
+
+        implicit val document: Document = Jsoup.parse(htmlFormat.body)
+
+        titleCheck(userScenario.specificExpectedResults.get.expectedTitle, userScenario.isWelsh)
+        h1Check(userScenario.specificExpectedResults.get.expectedHeading)
+        captionCheck(expectedCaption)
+        textOnPageCheck(userScenario.specificExpectedResults.get.youCanUseText, youCanUseTextSelector)
+        radioButtonCheck(yesText, radioNumber = 1, checked = false)
+        radioButtonCheck(noText, radioNumber = 2, checked = true)
+        buttonCheck(continueText, continueButtonSelector)
+        formPostLinkCheck(AssetsBenefitsController.submit(taxYearEOY, employmentId).url, formSelector)
+        welshToggleCheck(userScenario.isWelsh)
+      }
+
+      "return an error when a user submits an empty form" which {
+        implicit val authRequest: AuthorisationRequest[AnyContent] = getAuthRequest(userScenario.isAgent)
+        implicit val messages: Messages = getMessages(userScenario.isWelsh)
+
+        val htmlFormat = underTest(form(userScenario.isAgent).bind(Map(YesNoForm.yesNo -> "")), taxYearEOY, employmentId)
+
+        implicit val document: Document = Jsoup.parse(htmlFormat.body)
+
+        titleCheck(userScenario.specificExpectedResults.get.expectedErrorTitle, userScenario.isWelsh)
+        h1Check(userScenario.specificExpectedResults.get.expectedHeading)
+        captionCheck(expectedCaption)
+        textOnPageCheck(userScenario.specificExpectedResults.get.youCanUseText, youCanUseTextSelector)
+        radioButtonCheck(yesText, radioNumber = 1, checked = false)
+        radioButtonCheck(noText, radioNumber = 2, checked = false)
+        buttonCheck(continueText, continueButtonSelector)
+        formPostLinkCheck(AssetsBenefitsController.submit(taxYearEOY, employmentId).url, formSelector)
+
+        errorSummaryCheck(userScenario.specificExpectedResults.get.expectedErrorText, yesSelector)
+        errorAboveElementCheck(userScenario.specificExpectedResults.get.expectedErrorText, Some("value"))
+      }
+    }
+  }
+}
